refactor: consolidate duplicated utilities into core/parse-utils.js
Moved 4 runtime-neutral utility functions from worker-utils.js to core/parse-utils.js:
- getCanonicalMediaTimestamp
- sortMediaData
- normalizeSitePath
- getContentPathFromSitePath

Updated core/utils.js and core/paths.js to re-export these functions for backward compatibility.
Deleted worker-utils.js as it's no longer needed.
